refactor(workspace): extract actor handlers (w5b)
Extract ScanHandler, FileChangeHandler, and DocumentHandler from the workspace actor while keeping Actor as the coordinator. Add companion smoke tests for each handler and keep the event loop delegating through Actor wrappers so the branch stays green before later wiring work.
